+++ wifipumpkin3/core/servers/proxy/captiveflask.py
@@ -164,10 +164,10 @@ class CaptivePortal(ProxyMode):
     def parser_set_captiveflask(self, status, plugin_name):
         if len(plugin_name.split(".")) == 2:
             try:
-                # plugin_name = pumpkinproxy.no-cache
+                # plugin_name = captiveflask.FlaskDemo.En true
                 name_plugin, key_plugin = (
                     plugin_name.split(".")[0],
-                    plugin_name.split(".")[1],
+                    plugin_name.split(".")[1].split()[0],
                 )
                 if key_plugin in self.config.get_all_childname("plugins"):
                     self.setPluginActivated(key_plugin, status)
@@ -183,10 +183,10 @@ class CaptivePortal(ProxyMode):
                 print(display_messages("unknown sintax command", error=True))
         elif len(plugin_name.split(".")) == 3:
             try:
-                # plugin_name = pumpkinproxy.beef.url_hook
+                # plugin_name = captiveflask.FlaskDemo.En true
                 name_plugin, key_plugin = (
                     plugin_name.split(".")[1],
-                    plugin_name.split(".")[2],
+                    plugin_name.split(".")[2].split()[0],
                 )
                 if key_plugin in self.config.get_all_childname(
                     "set_{}".format(name_plugin)