+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 ### Removed
 
 ### Fixed
+- fixed bug unknow plugin when try to use command `set captiveflask` thanks @MrFabNc 
 - fixed set command for settings sniffkin3, pumpkinproxy, security
 - fixed ImportError: cannot import name 'json' from 'itsdangerous' 
 - fixed markupsafe==2.0.1 tos solve deprecated the JSON API.